split up service tests
